+++ streams/src/commonMain/kotlin/com/mirego/trikot/streams/reactive/MutablePublisher.kt
@@ -1,0 +1,9 @@
+package com.mirego.trikot.streams.reactive
+
+import org.reactivestreams.Publisher
+
+interface MutablePublisher<T> : Publisher<T> {
+    var value: T?
+    var error: Throwable?
+    fun complete()
+}
